Script launching an OBS streaming virtual machine
My laptop is a sadtop, and can't zoom + OBS at the same time.

This pushes the resonsibility for compositing the video and audio
sources when streaming to Leviathan.

I _may_ want to consider switching my GPU passthrough from the mac to
the streaming compositor, or add a dedicated GPU + Capture Card.