Upgrade to Ubuntu 20.04 LTS and stop logging in as root
It seems like being on the LTS release of Ubuntu will make things a bit
easier; but this could be a horrible idea. Also, I noticed that I was
logging in as root all over the place, instead of sshing in as me and
using `sudo`.

This also makes it possible to use the `-u` flag to specify the username
being logged in as instead of assuming that we always want to ssh in as
root.