Configure docker for use by workstation guests
Previously, I had installed docker by selecting it as part of the server
installation. However, that leaves us on a version of docker that may be
old.

Automating the installation of docker on both the host and the guests
ensures that we're using a consistent version of docker for
containerization, as well as ensures I don't have to remember what to do
when I rebuild the Leviathan as a workstation-host again.