🛠📝 Sprout remote management tasks and document them
As I find bits and bobs that enable or support remote management I'm
grouping them into the `maintenance/remote-management.yml` file. I can
imagine adding packages that are necessary for ansible to do it's thing
to be added here.

I'm trying to think about how to disambiguate between this space being
used for enabling the machine to be managed remotely _vs_ enabling this
machine to _perform_ remote management.