🛠 Use pip to install ansible instead of apt
This may turn out to be a very bad idea, but I'm not entirely sure how
to make ansible play nicely with python-as-installed-through-asdf vs
python-as-installed-by-ubuntu.

It _seems like_ what I really want is to always use the system-level
python and _never_ use the python discovered on path; but I need to
learn more about how to make that work.